requirements.txt
Flask>=3.0
requests>=2.31
pandas>=2.1
python-dotenv>=1.0
matplotlib>=3.8